ingetion/requirements_fixed.txt
# === Extraction de documents ===
pymupdf>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
openpyxl>=3.1.2
Pillow>=10.2.0
pytesseract>=0.3.10

# === Utilitaires ===
loguru>=0.7.2
python-dotenv>=1.0.0
watchdog>=4.0.0
beautifulsoup4>=4.12.3
markdownify>=0.11.6
tabulate>=0.9.0
requests>=2.31.0
openai>=1.0.0

# === API ===
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9

# === ChromaDB ===
chromadb>=0.4.22

# === LLM / LangChain (versions compatibles entre elles) ===
groq==0.5.0
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langgraph==0.1.19
langchain-groq==0.1.9
